Allow any Promises/A+ implementation
features
 * Use global.Promise or options.Promise instead of Q
 * Expose Promise on returned Handlebars object
tests
 * Upgrade tests as discussed in #20
 * Add tests for bluebird, es6-promise, and global.Promise
 * Pass created Handlebars engine to tests via mocha context
docs
 * Update examples & readme
