+++ examples/setup-with-global-bluebird.js
@@ -1,11 +1,14 @@
+// -----<snip>-------
+global.Promise = require('bluebird')
 var promisedHandlebars = require('../')
-var Q = require('q')
-var Handlebars = promisedHandlebars(require('handlebars'), { Promise: Q.Promise })
+// By default promisedHandlebars uses global.Promise,
+var Handlebars = promisedHandlebars(require('handlebars'))
+// -----</snip>-------
 
 // Register a helper that returns a promise
 // Helpers do not have to return a promise of the sane
 Handlebars.registerHelper('helper', function (value) {
-  return Q.delay(100).then(function () {
+  return Promise.delay(100).then(function () {
     return value
   })
 })
@@ -16,4 +19,4 @@ var template = Handlebars.compile('123{{helper a}}456{{helper b}}')
 template({
   a: 'abc',
   b: 'xyz'
-}).done(console.log)
+}).then(console.log)
